+++ components/shared/include/shared.h
@@ -6,6 +6,7 @@ typedef struct {
     float temperature;
     float humidity;
     float pressure;
+    uint16_t lux;
 } shared_data_t;
 
 extern shared_data_t shared_data;