+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 This changelog suppose to follow rules defined in the [changelog.md](https://changelog.md)
 
-## 0.2.0 - TBD
+## 0.2.0 - 2024-06-18
+
+- **Added**: VCNL4040 support (experimental)
+- **Changed**: `VOGON` device name to `VOG1`
 
 ## 0.1.0 - 2024-06-10